Refactoring code.

+++ PayItForward.Data.Models/User.cs
@@ -33,7 +33,7 @@ namespace PayItForward.Data.Models
         [StringLength(20)]
         public string LastName { get; set; }
 
-        public double AvilableMoneyAmount { get; set; }
+        public decimal AvilableMoneyAmount { get; set; }
 
         [Column(TypeName = "varchar(200)")]
         public string AvatarUrl { get; set; }
